move tests to test file
